Back/Code/src/main/java/com/example/acdc/repository/UserRepository.java
package com.example.acdc.repository;

import com.example.acdc.domain.User;
import jakarta.persistence.EntityManager;
import jakarta.persistence.PersistenceContext;
import org.springframework.stereotype.Repository;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@Repository
@Transactional(readOnly = true)
public class UserRepository {

    @PersistenceContext
    private EntityManager em;

    @Transactional
    public void save(User user) {
        em.persist(user);
    }

    public User findOne(Long id) {
        return em.find(User.class, id);
    }

    public List<User> findByName(String name) {
        return em.createQuery("select U from User as U where U.name = :name", User.class)
                .setParameter("name", name)
                .getResultList();
    }
}
